VelinWebsite-GitHub-Upload/requirements.txt
Django==6.0.3
Pillow==12.2.0
dj-database-url==3.1.2
psycopg[binary]==3.3.3
django-allauth[socialaccount]==65.15.1
